Replace parseInt -> parseFloat for margin-based examples (#291)

+++ src/comparisons/elements/outer_height_with_margin/ie8.js
@@ -2,7 +2,7 @@ function outerHeight(el) {
   var height = el.offsetHeight;
   var style = el.currentStyle || getComputedStyle(el);
 
-  height += parseInt(style.marginTop) + parseInt(style.marginBottom);
+  height += parseFloat(style.marginTop) + parseFloat(style.marginBottom);
   return height;
 }
